move opt-out to "danger zone" in "advanced" tab
